requirements.txt
openai>=1.0.0
streamlit>=1.0.0
python-dotenv>=0.19.0
requests>=2.26.0
Pillow>=9.0.0
pydub==0.25.1
elevenlabs>=0.2.0
beautifulsoup4>=4.9.3
python-magic>=0.4.27 